Added an overview field on metric type.
Changed the metric page template so that the javascript filename becomes
the id of a metrics div.

The metric page template now uses a view method to find all CSS files
contained within the metric and add a link element to them. This uses
link within the body of the html which could cause problems down the
road.